pyproject.toml
[tool.poetry]
name = "openai-code-interpreter-demo"
version = "0.1.0"
description = ""
authors = ["seiya303"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.23.3"
streamlit = "^1.33.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
